refactor: move balances loading to use useBalancesQuery
