+++ dapp/src/constants/env.js
@@ -1,0 +1,2 @@
+export const isDevelopment = process.env.NODE_ENV === 'development'
+export const isProduction = process.env.NODE_ENV === 'production'
