+++ dapp/src/constants/queryKeys.js
@@ -1,3 +1,4 @@
 export const QUERY_KEYS = {
   TransactionHistory: (account) => ['transactionHistory', { account }],
+  Balances: (account) => ['balances', { account }],
 }
