+++ dapp/src/utils/analytics.js
@@ -1,10 +1,9 @@
 import Analytics from 'analytics'
 import googleAnalytics from '@analytics/google-analytics'
 import mixpanel from '@analytics/mixpanel'
+import { isProduction, isDevelopment } from 'constants/env'
 
 const MIXPANEL_ID = process.env.MIXPANEL_ID
-const isProduction = process.env.NODE_ENV === 'production'
-const isDevelopment = process.env.NODE_ENV === 'development'
 const isStaging = process.env.STAGING === 'true'
 
 let mixpanelId = MIXPANEL_ID || 'dev_token'
